Squashed commit of the following:
commit 62e4744cd022c110c5acb507dedaf08ab43fb361
Author: Rasmus Krämer <git@rfk.io>
Date:   Sat Jun 1 15:00:20 2024 +0200

commit 57e1a0982ade3fbf5332d335593fadb415b448ca
Author: Rasmus Krämer <git@rfk.io>
Date:   Fri May 31 10:59:50 2024 +0200

commit 5f713c4d4e86445cf7d08fece7c4840bc9f9c479
Author: Rasmus Krämer <git@rfk.io>
Date:   Thu May 30 18:18:46 2024 +0200

commit c22a979e779f14f17fa92df36ba1eb96b8049e51
Author: Rasmus Krämer <git@rfk.io>
Date:   Tue May 28 17:36:46 2024 +0200

commit 732dacb3b8f4be37454effcc2563be8b8ead5fc7
Author: Rasmus Krämer <git@rfk.io>
Date:   Tue May 28 10:01:17 2024 +0200

commit 620fa3e3e7cd7adff1690d9fff88c29b3e6e95af
Author: Rasmus Krämer <git@rfk.io>
Date:   Mon May 27 18:18:03 2024 +0200

commit 14a6747766a2e3850afac12664fe0d25bcbaf459
Author: Rasmus Krämer <git@rfk.io>
Date:   Sat May 25 00:22:15 2024 +0200

commit 254a883fd482413dbd233c1c6869c71f014d9237
Author: Rasmus Krämer <git@rfk.io>
Date:   Fri May 24 20:01:04 2024 +0200

commit 28cae49845229daf7b331cb7128af6dc0aa16170
Author: Rasmus Krämer <git@rfk.io>
Date:   Thu May 23 17:35:21 2024 +0200

commit 34f610c41a2bc8e913937136965027fbd2ade1ea
Author: Rasmus Krämer <git@rfk.io>
Date:   Thu May 23 12:30:53 2024 +0200

commit 151b81c496211cc93389b120092a16735a19149c
Author: Rasmus Krämer <git@rfk.io>
Date:   Thu May 23 00:32:08 2024 +0200

commit 974a7e5bd225f5dbda1b7d946baaa29ac9674467
Author: Rasmus Krämer <git@rfk.io>
Date:   Wed May 22 23:51:38 2024 +0200

commit 261e4dc344f36868bdd03529e55f1845d103d2f4
Author: Rasmus Krämer <git@rfk.io>
Date:   Wed May 22 16:29:10 2024 +0200

commit 14522424e8b9f65327fc129b92aaaad9335dbeb3
Author: Rasmus Krämer <git@rfk.io>
Date:   Wed May 22 12:00:06 2024 +0200

commit 9b64590e292a3cd1e0607e7a863957c364f642dd
Author: Rasmus Krämer <git@rfk.io>
Date:   Tue May 14 18:47:05 2024 +0200

+++ AmpFinKit/Sources/AFOffline/OfflineManager/OfflineManager+Cleanup.swift
@@ -7,82 +7,78 @@
 
 import Foundation
 import SwiftData
-import AFBase
-
-// MARK: Remove tracks
+import AFFoundation
+import AFNetwork
 
 public extension OfflineManager {
     @MainActor
-    func deleteAll() throws {
-        for album in try getOfflineAlbums() {
+    func delete() throws {
+        for album in try offlineAlbums() {
             try delete(offlineAlbum: album)
         }
         
-        for playlist in try getOfflinePlaylists() {
+        for playlist in try offlinePlaylists() {
             try delete(playlist: playlist)
         }
         
-        for track in try getOfflineTracks() {
+        for track in try offlineTracks() {
             delete(track: track)
         }
         
         try DownloadManager.shared.cleanupDirectory()
     }
     
+    @MainActor
     func syncPlaysToJellyfinServer() {
-        Task.detached { @MainActor in
-            let plays = try PersistenceManager.shared.modelContainer.mainContext.fetch(FetchDescriptor<OfflinePlay>())
-            
-            Self.logger.info("Syncing \(plays.count) plays to the server")
-            
+        guard let plays = try? PersistenceManager.shared.modelContainer.mainContext.fetch(FetchDescriptor<OfflinePlay>()) else {
+            return
+        }
+        
+        Self.logger.info("Syncing \(plays.count) plays to the server")
+        
+        Task.detached {
             for play in plays {
-                Task.detached {
-                    do {
-                        try await JellyfinClient.shared.reportPlaybackStopped(trackId: play.trackId, positionSeconds: play.positionSeconds)
-                        
-                        Task.detached { @MainActor in
-                            PersistenceManager.shared.modelContainer.mainContext.delete(play)
-                        }
-                    } catch {
-                        Self.logger.fault("Error while syncing play to Jellyfin server \(play.trackId) (\(play.positionSeconds)")
-                    }
+                try await JellyfinClient.shared.playbackStopped(identifier: play.trackId, positionSeconds: play.positionSeconds)
+                
+                await MainActor.run {
+                    PersistenceManager.shared.modelContainer.mainContext.delete(play)
                 }
             }
         }
     }
     
-    func updateOfflineItems() {
+    func update() {
         let waitTime: Double = 60 * 60 * 12
-        let lastDonation = UserDefaults.standard.double(forKey: "lastOfflineItemUpdate")
+        let lastUpdate = UserDefaults.standard.double(forKey: "lastOfflineItemUpdate")
         
-        if lastDonation + waitTime > Date.timeIntervalSinceReferenceDate {
-            updateOfflineFavorites()
+        if lastUpdate + waitTime > Date.timeIntervalSinceReferenceDate {
+            updateFavorites()
         } else {
             UserDefaults.standard.set(Date.timeIntervalSinceReferenceDate, forKey: "lastOfflineItemUpdate")
             
-            Task.detached { @MainActor in
-                for album in try getAlbums() {
+            Task {
+                for album in try await albums() {
                     try await download(album: album)
                 }
                 
-                for playlist in try getPlaylists() {
+                for playlist in try await playlists() {
                     try await download(playlist: playlist)
                 }
                 
-                try? removeOrphanedTracks()
+                try? await removeOrphans()
             }
         }
     }
     
     @MainActor
-    func removeOrphanedTracks() throws {
-        let albums = try getOfflineAlbums()
-        let playlists = try getOfflinePlaylists()
+    func removeOrphans() throws {
+        let albums = try offlineAlbums()
+        let playlists = try offlinePlaylists()
         
-        let albumTrackIds = reduceToChildrenIds(parents: albums)
-        let playlistTrackIds = reduceToChildrenIds(parents: playlists)
+        let albumTrackIds = reduceToChildrenIdentifiers(parents: albums)
+        let playlistTrackIds = reduceToChildrenIdentifiers(parents: playlists)
         
-        let tracks = try getOfflineTracks()
+        let tracks = try offlineTracks()
         let orphaned = tracks.filter { !albumTrackIds.contains($0.id) && !playlistTrackIds.contains($0.id) }
         
         for orphan in orphaned {
